us: add old Midwest Wireless SIDs (bought by Alltel in 2005)
And of course Alltel was bought by Verizon in 2008/2009 and some
markets went to AT&T and some to ATN, and most to Verizon.  We'll
have to sort out the SID list at some point for that based on what
markets Verizon had to divest to AT&T and ATN, the rest to Verizon.
ATN is doing business as Alltel in 22 markets, but the markets that
were kept by Verizon have been fully converted to Verizon branding
and service.