dtd: add initial attach apn type
This patch adds the "ia" apn  type to be able to provision 4G

LTE has the concept of the Initial Attach APN for the
initial PDP context. This initial PDP context is automatically
attached once registered on an LTE network.

oFono currently supports setting the Initial Attach APN but it
has no way of provisioning this value.

https://bugzilla.gnome.org/show_bug.cgi?id=775985