italy: add Vodafone Internet Key
Vodafone data SIMs need to use the web.omnitel.it APN, otherwise
you connect as a smartphone and are billed according to a completely
different plan.  The DNS of "Internet Facile (old)" however does not
work.

Note that both APNs are necessary because if you have a smartphone
SIM you will need to use mobile.vodafone.it or again the billing
will be wrong.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>

https://bugzilla.gnome.org/show_bug.cgi?id=776355