Expand DTD to include information for Visual Voicemail
This commit expands the DTD include settings for visual voicemail (VVM) activation. The two document references are:

[1] https://www.gsma.com/newsroom/wp-content/uploads/2012/07/OMTP_VVM_Specification_1_3.pdf
This is the primary VVM specification

[2] https://source.android.com/devices/tech/config/voicemail
This includes some information that is not obvious from the VVM Specification.

VVM's control signals are over Short Message Service (SMS), and there are (at least) three required settings:
- The VVM Specification ("otmp" is the primary one [1], but [2] notes that there is a second one, "cvvm", and T-Mobile USA is at least one carrier that uses "cvvm"
- The destination visual voicemail number that communicates with the client, and
- The Carrier Prefix, which is the prefix to every VVM control SMS.

With these settings, a VVM client can activate and set up visual voicemail, as all other required settings will be received from the carrier (and thus are not needed in mobile-broadband-provider-info). One such client in development that will use these settings is vvmd:
https://gitlab.com/kop316/vvmd/

To my knowledge, all carriers that have VVM adhere to the VVM OMTP/CVVM standard (as these are the only standards that Android has), and will use settings.