us: add US Mobile (T-Mobile MVNO): mcc="310" mnc="260"
This patch adds US Mobile (T-Mobile / Verizon MVNO) with the T-Mobile USA
gsm APN settings. The type of "internet" is used to allow a single combined
Cellular Data / MMS APN entry. Note that 310/260 is already specified in
the T-Mobile provider entry as well as the SpeedTalk Mobile provider entry
but these are the right network settings for US Mobile's T-Mobile network
entry.

https://gitlab.gnome.org/GNOME/mobile-broadband-provider-info/-/merge_requests/54