dtd: allow "mms-internet-hipri" and "mms-internet-hipri-fota" types
Update to the current AT&T APNs requires specifying multiple usages.

I added the possibility of using a chain of usages in one usage i.e.
usage="<usage1>-<usage2>-...". It's not a perfect system, but it will work
until someone needs a larger number of combinations of usages, as the ones
I use are hard-coded.

https://gitlab.gnome.org/GNOME/mobile-broadband-provider-info/-/merge_requests/81