serviceproviders: Allow to add emergency numbers
Emergency numbers are needed e.g. in call applications to figure out the
valid numbers based on the country the device is in.

Hence allow to add that information to the country block of the
serviceprovider XML.

We only want to list numbers that are free to dial and that can work
when the SIM is missing / locked.

Refrence is ETSI TS 122 101 V14.8.0 Section 10

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/GNOME/mobile-broadband-provider-info/-/merge_requests/126>