serviceproviders: Allow to add Cell Broadcast information
The list of channels is needed to be able to configure any modems to
receive messages at the appropriate levels and also for user space to
figure out how important a received message is (so it can show a warning
with a proper severity / importance). We want to add language
information at a later point so UEs can allow for language selection.

This adds the necessary schema bits.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/GNOME/mobile-broadband-provider-info/-/merge_requests/127>